Modify: Accordion 컴포넌트 적용 및 관련 수정사항
- index.html에 Accordion 컴포넌트를 적용했습니다.
- style.css에 qna-template.css로 되어있던걸 accordion-template.css로 수정했습니다.
- src/ 폴더에 있던 accordion-components.css를 src/javascripts/폴더로 이동하였습니다.
- accordion-components.js의 DOMLoad 이벤트리스너 부분을 제거했습니다.
- accordion-components.html/accordion-template.html의 스크립트 로드 부분을 수정했습니다.

resolves #62
