+++ index.html
@@ -131,6 +131,84 @@
         </div>
       </section>
       <!-- qna -->
+      <section id="accordion-components-wrap" role="region" aria-labelledby="faq-title">
+        <h3 id="faq-title" class="accordion-title">자주 하는 질문</h3>
+        <div class="accordion-components" role="group">
+          <button type="button" class="accordion-head" aria-expanded="false" aria-controls="faq1">미터기 요금을 중복으로 결제하였습니다 (앱+현장)</button>
+          <div id="faq1" class="accordion-body" hidden>
+            <div class="accordion-body-inner">
+              <p>비슷한 금액의 요금이 중복 청구 된 경우 아래와 같이 생각하실 수 있습니다.</p>
+              <ol>
+                <li>가족 또는 지인이 동일한 카드를 등록하여 사용하지 않았는지 확인해보세요.</li>
+                <li>우연히 같은 요금의 다른 차량 서비스를 이용하지 않았는지 확인해보는 것이 좋습니다.</li>
+                <li>어떠한 이유로 잘못 청구된 여정 요금이 변경된 경우가 없는지 확인해 보세요. 요금이 변경되면 업데이트 된 영수증이 발급됩니다.</li>
+              </ol>
+              <p>앱 결제 차량에서 실수로 현장에서 요금을 지불하여 중복 결제가 된 경우 결제된 카드 내역 또는 영수증 캡처하여 아래 링크를 통해 알려주세요.</p>
+              <strong>[고객지원 > 여정 및 요금 > 내 요금 및 수수료 검토 > 미터기 요금을 중복으로 결제하였습니다]</strong>
+            </div>
+          </div>
+        </div>
+        <div class="accordion-components" role="group">
+          <button type="button" class="accordion-head" aria-expanded="false" aria-controls="faq2">미터기 요금 외에 다른 요금이 결제 됐습니다. (통행료, 가승인 등)</button>
+          <div id="faq2" class="accordion-body" hidden>
+            <div class="accordion-body-inner">
+              <p>비슷한 금액의 요금이 중복 청구 된 경우 아래와 같이 생각하실 수 있습니다.</p>
+              <ol>
+                <li>가족 또는 지인이 동일한 카드를 등록하여 사용하지 않았는지 확인해보세요.</li>
+                <li>우연히 같은 요금의 다른 차량 서비스를 이용하지 않았는지 확인해보는 것이 좋습니다.</li>
+                <li>어떠한 이유로 잘못 청구된 여정 요금이 변경된 경우가 없는지 확인해 보세요. 요금이 변경되면 업데이트 된 영수증이 발급됩니다.</li>
+              </ol>
+              <p>앱 결제 차량에서 실수로 현장에서 요금을 지불하여 중복 결제가 된 경우 결제된 카드 내역 또는 영수증 캡처하여 아래 링크를 통해 알려주세요.</p>
+              <strong>[고객지원 > 여정 및 요금 > 내 요금 및 수수료 검토 > 미터기 요금을 중복으로 결제하였습니다]</strong>
+            </div>
+          </div>
+        </div>
+        <div class="accordion-components" role="group">
+          <button type="button" class="accordion-head" aria-expanded="false" aria-controls="faq3">분실물과 관련해 기사에게 연락하기</button>
+          <div id="faq3" class="accordion-body" hidden>
+            <div class="accordion-body-inner">
+              <p>비슷한 금액의 요금이 중복 청구 된 경우 아래와 같이 생각하실 수 있습니다.</p>
+              <ol>
+                <li>가족 또는 지인이 동일한 카드를 등록하여 사용하지 않았는지 확인해보세요.</li>
+                <li>우연히 같은 요금의 다른 차량 서비스를 이용하지 않았는지 확인해보는 것이 좋습니다.</li>
+                <li>어떠한 이유로 잘못 청구된 여정 요금이 변경된 경우가 없는지 확인해 보세요. 요금이 변경되면 업데이트 된 영수증이 발급됩니다.</li>
+              </ol>
+              <p>앱 결제 차량에서 실수로 현장에서 요금을 지불하여 중복 결제가 된 경우 결제된 카드 내역 또는 영수증 캡처하여 아래 링크를 통해 알려주세요.</p>
+              <strong>[고객지원 > 여정 및 요금 > 내 요금 및 수수료 검토 > 미터기 요금을 중복으로 결제하였습니다]</strong>
+            </div>
+          </div>
+        </div>
+        <div class="accordion-components" role="group">
+          <button type="button" class="accordion-head" aria-expanded="false" aria-controls="faq4">택시가 잡혀 기다렸으나 저와는 반대방향으로 이동했어요.</button>
+          <div id="faq4" class="accordion-body" hidden>
+            <div class="accordion-body-inner">
+              <p>비슷한 금액의 요금이 중복 청구 된 경우 아래와 같이 생각하실 수 있습니다.</p>
+              <ol>
+                <li>가족 또는 지인이 동일한 카드를 등록하여 사용하지 않았는지 확인해보세요.</li>
+                <li>우연히 같은 요금의 다른 차량 서비스를 이용하지 않았는지 확인해보는 것이 좋습니다.</li>
+                <li>어떠한 이유로 잘못 청구된 여정 요금이 변경된 경우가 없는지 확인해 보세요. 요금이 변경되면 업데이트 된 영수증이 발급됩니다.</li>
+              </ol>
+              <p>앱 결제 차량에서 실수로 현장에서 요금을 지불하여 중복 결제가 된 경우 결제된 카드 내역 또는 영수증 캡처하여 아래 링크를 통해 알려주세요.</p>
+              <strong>[고객지원 > 여정 및 요금 > 내 요금 및 수수료 검토 > 미터기 요금을 중복으로 결제하였습니다]</strong>
+            </div>
+          </div>
+        </div>
+        <div class="accordion-components" role="group">
+          <button type="button" class="accordion-head" aria-expanded="false" aria-controls="faq5">프로모션 코드가 적용되지 않았습니다.</button>
+          <div id="faq5" class="accordion-body" hidden>
+            <div class="accordion-body-inner">
+              <p>비슷한 금액의 요금이 중복 청구 된 경우 아래와 같이 생각하실 수 있습니다.</p>
+              <ol>
+                <li>가족 또는 지인이 동일한 카드를 등록하여 사용하지 않았는지 확인해보세요.</li>
+                <li>우연히 같은 요금의 다른 차량 서비스를 이용하지 않았는지 확인해보는 것이 좋습니다.</li>
+                <li>어떠한 이유로 잘못 청구된 여정 요금이 변경된 경우가 없는지 확인해 보세요. 요금이 변경되면 업데이트 된 영수증이 발급됩니다.</li>
+              </ol>
+              <p>앱 결제 차량에서 실수로 현장에서 요금을 지불하여 중복 결제가 된 경우 결제된 카드 내역 또는 영수증 캡처하여 아래 링크를 통해 알려주세요.</p>
+              <strong>[고객지원 > 여정 및 요금 > 내 요금 및 수수료 검토 > 미터기 요금을 중복으로 결제하였습니다]</strong>
+            </div>
+          </div>
+        </div>
+      </section>
       <!-- apply -->
       <section class="apply-template">
         <h1 class="apply-header">UT로 운행하기</h1>
