+++ src/accordion-components.html
@@ -6,8 +6,7 @@
     <title>아코디언 컴포넌트</title>
     <link rel="stylesheet" href="./styles/components/accordion-components.css" />
     <link rel="stylesheet" href="./style.css" />
-    <script type="module" src="./main.js"></script>
-    <script type="module" src="./accordion-components.js"></script>
+    <script type="module" src="/src/javascripts/accordion-components.js"></script>
   </head>
   <body>
     <section id="accordion-components-wrap">
