+++ src/accordion-template.html
@@ -5,9 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>아코디언 컴포넌트</title>
     <link rel="stylesheet" href="./styles/components/accordion-components.css" />
-    <link rel="stylesheet" href="./style.css" />
-    <script type="module" src="./main.js"></script>
-    <script type="module" src="./accordion-components.js"></script>
+    <link rel="stylesheet" href="/src/style.css" />
+    <script type="module" src="/src/javascripts/accordion-components.js"></script>
   </head>
   <body>
     <section id="accordion-components-wrap" role="region" aria-labelledby="faq-title">
