+++ src/main.js
@@ -1,1 +1,2 @@
 import './javascripts/input-component.js';
+import './javascripts/accordion-components.js';
